Lift loggedIn state and all related logic up to App
