+++ src/App.js
@@ -3,6 +3,7 @@ import { Normalize } from 'styled-normalize';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 
 import GlobalStyle from './globalStyle';
+import { findSessionCookie } from './util';
 import {
   AdminRequestsPage,
   AdminReservationsPage,
@@ -11,13 +12,44 @@ import {
 import { CitizensPage } from './UI/citizens';
 
 class App extends Component {
+  state = {
+    loggedIn: false
+  };
+
+  componentDidMount() {
+    this.detectActiveSession();
+  }
+
+  detectActiveSession = () => {
+    const isSessionCookie = findSessionCookie();
+    this.authenticateAdmin(isSessionCookie);
+  };
+
+  authenticateAdmin = loggedIn => {
+    this.setState({
+      loggedIn
+    });
+  };
+
   render() {
+    const { loggedIn } = this.state;
+
     return (
       <Router>
         <Normalize />
         <GlobalStyle />
         <Switch>
-          <Route path="/" exact component={CitizensPage} />
+          <Route
+            path="/"
+            exact
+            render={props => (
+              <CitizensPage
+                {...props}
+                loggedIn={loggedIn}
+                authenticateAdmin={auth => this.authenticateAdmin(auth)}
+              />
+            )}
+          />
           <Route path="/admin-requests" component={AdminRequestsPage} />
           {/* <Route path="/admin-reservations" component={AdminReservationsPage} />
         <Route path="/admin-halls" component={AdminHallsPage} /> */}
